# perf: use fast unsafe bytes->string convertion