--- a/unsafe.go
+++ b/unsafe.go
@@ -1,0 +1,8 @@
+package iavl
+
+import ibytes "github.com/cosmos/iavl/internal/bytes"
+
+var (
+	unsafeToStr = ibytes.UnsafeBytesToStr
+	unsafeToBz  = ibytes.UnsafeStrToBytes
+)