--- a/unsaved_fast_iterator.go
+++ b/unsaved_fast_iterator.go
@@ -6,6 +6,8 @@ import (
 	"sort"
 
 	dbm "github.com/tendermint/tm-db"
+
+	ibytes "github.com/cosmos/iavl/internal/bytes"
 )
 
 var (
@@ -18,29 +20,19 @@ var (
 // it iterates over the latest state via fast nodes,
 // taking advantage of keys being located in sequence in the underlying database.
 type UnsavedFastIterator struct {
-	start, end []byte
-
-	valid bool
-
-	ascending bool
-
-	err error
-
-	ndb *nodeDB
-
-	unsavedFastNodeAdditions map[string]*FastNode
-
-	unsavedFastNodeRemovals map[string]interface{}
-
-	unsavedFastNodesToSort []string
-
-	nextKey []byte
-
-	nextVal []byte
-
-	nextUnsavedNodeIdx int
-
+	start, end   []byte
+	valid        bool
+	ascending    bool
+	err          error
+	ndb          *nodeDB
+	nextKey      []byte
+	nextVal      []byte
 	fastIterator dbm.Iterator
+
+	nextUnsavedNodeIdx       int
+	unsavedFastNodeAdditions map[string]*FastNode
+	unsavedFastNodeRemovals  map[string]interface{}
+	unsavedFastNodesToSort   []string
 }
 
 var _ dbm.Iterator = (*UnsavedFastIterator)(nil)
@@ -71,7 +63,7 @@ func NewUnsavedFastIterator(start, end []byte, ascending bool, ndb *nodeDB, unsa
 			continue
 		}
 
-		iter.unsavedFastNodesToSort = append(iter.unsavedFastNodesToSort, string(fastNode.key))
+		iter.unsavedFastNodesToSort = append(iter.unsavedFastNodesToSort, ibytes.UnsafeBytesToStr(fastNode.key))
 	}
 
 	sort.Slice(iter.unsavedFastNodesToSort, func(i, j int) bool {
@@ -142,8 +134,8 @@ func (iter *UnsavedFastIterator) Next() {
 		return
 	}
 
+	diskKeyStr := ibytes.UnsafeBytesToStr(iter.fastIterator.Key())
 	if iter.fastIterator.Valid() && iter.nextUnsavedNodeIdx < len(iter.unsavedFastNodesToSort) {
-		diskKeyStr := string(iter.fastIterator.Key())
 
 		if iter.unsavedFastNodeRemovals[diskKeyStr] != nil {
 			// If next fast node from disk is to be removed, skip it.
@@ -186,7 +178,7 @@ func (iter *UnsavedFastIterator) Next() {
 
 	// if only nodes on disk are left, we return them
 	if iter.fastIterator.Valid() {
-		if iter.unsavedFastNodeRemovals[string(iter.fastIterator.Key())] != nil {
+		if iter.unsavedFastNodeRemovals[diskKeyStr] != nil {
 			// If next fast node from disk is to be removed, skip it.
 			iter.fastIterator.Next()
 			iter.Next()