Avoid unnecessary allocation by using fast unsafe bytes -> string conversion

Notes:
+ I didn't add conversions to the tests - I'm planning other PR there to cleanup key / values repetitions.