@robert-zaremba 2022-07-25
Benchmark:
```
benchmark                                old ns/op     new ns/op     delta
BenchmarkGetNonMembership/fast-16        143368        129285        -9.82%
BenchmarkGetNonMembership/regular-16     10305208      8466036       -17.85%
BenchmarkImmutableAvlTreeMemDB-16        150397        139207        -7.44%

benchmark                             old allocs     new allocs     delta
BenchmarkImmutableAvlTreeMemDB-16     437            392            -10.30%

benchmark                             old bytes     new bytes     delta
BenchmarkImmutableAvlTreeMemDB-16     33513         32100         -4.22%
```