@robert-zaremba 2022-07-25
turns out that it's significant optimization