@robert-zaremba 2022-08-03
After removing the fast convertion, there is a slight performance degradation:
* before = with unsafe string in cache.go maps
* after = with normal `string(key)` lookup in cache.go maps.

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkImmutableAvlTreeMemDB-16     138184        140417        +1.62%

benchmark                             old allocs     new allocs     delta
BenchmarkImmutableAvlTreeMemDB-16     392            419            +6.89%

benchmark                             old bytes     new bytes     delta
BenchmarkImmutableAvlTreeMemDB-16     32154         32928         +2.41%
```

I repeated it few times and the memory allocation results were similar (there was more variation in the ns/op)